+++ changelog.md
@@ -2,6 +2,7 @@ Changelog
 
 ### #4 2017-11-11 - Kabir Cahill
 - added fleet edit page
+- added fleet wackyid dropdown
 
 ### #3 2017-11-11 - Kent Huang
 - added model for shared planes in the wacky server